+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Goro want chocolate".
 - Tests for "Photo booth transformation".
 - Tests for "Bulls and cows".
+- Tests for "CGFunge interpreter".
 
 ## [1.11.0] - 2022-07-31
 ### Added